requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
diffusers>=0.18.0
accelerate>=0.20.0

# Multi-modal Models
open-clip-torch>=2.20.0
sentence-transformers>=2.2.0
clip-by-openai>=1.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0
torchmetrics>=0.11.0
albumentations>=1.3.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
wandb>=0.15.0
tensorboard>=2.13.0

# Data Processing
nltk>=3.8
spacy>=3.6.0
jieba>=0.42.1
PyYAML>=6.0
h5py>=3.8.0

# Networking and APIs
requests>=2.31.0
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.22.0

# Configuration and Environment
python-dotenv>=1.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Testing and Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Performance and Optimization
numba>=0.57.0
cupy-cuda12x>=12.0.0; sys_platform != "darwin"
faiss-gpu>=1.7.4; sys_platform != "darwin"
faiss-cpu>=1.7.4; sys_platform == "darwin"

# Text Processing
ftfy>=6.1.0
regex>=2023.6.0
tqdm>=4.65.0

# Additional utilities
psutil>=5.9.0
colorama>=0.4.6
rich>=13.4.0